Added some tests for content mapping part
