+++ mapper/mapper.go
@@ -14,12 +14,6 @@ type AggregateCPHMapper struct {
 	cphValidator CPHValidator
 }
 
-type contentCPHMapper struct {
-}
-
-type complementaryContentCPHMapper struct {
-}
-
 func NewAggregateCPHMapper() *AggregateCPHMapper {
 	return &AggregateCPHMapper{cphValidator: NewDefaultCPHValidator(), cphMappers: []CPHMapper{&contentCPHMapper{}, &complementaryContentCPHMapper{}}}
 }
@@ -41,34 +35,4 @@ func (m *AggregateCPHMapper) MapContentPlaceholder(mpc *model.MethodeContentPlac
 		}
 	}
 	return transformedResults, nil
-}
-
-func (cm *contentCPHMapper) MapContentPlaceholder(mcp *model.MethodeContentPlaceholder) ([]model.UppContent, *utility.MappingError) {
-	if mcp.IsInternalCPH() {
-		return []model.UppContent{}, nil
-	} else {
-		if mcp.Attributes.IsDeleted {
-			return []model.UppContent{model.NewUppContentPlaceholderDelete(mcp)}, nil
-		}
-
-		uppContent, err := model.NewUppContentPlaceholder(mcp)
-		if err != nil {
-			return nil, err
-		}
-
-		return []model.UppContent{uppContent}, nil
-	}
-}
-
-func (ccm *complementaryContentCPHMapper) MapContentPlaceholder(mcp *model.MethodeContentPlaceholder) ([]model.UppContent, *utility.MappingError) {
-	uuidToSet := mcp.UUID
-	if mcp.IsInternalCPH() {
-		uuidToSet = mcp.Attributes.LinkedArticleUUID
-	}
-
-	if mcp.Attributes.IsDeleted {
-		return []model.UppContent{model.NewUppComplementaryContentDelete(mcp, uuidToSet)}, nil
-	}
-
-	return []model.UppContent{model.NewUppComplementaryContent(mcp, uuidToSet)}, nil
 }
